// src/repository/openapi/model/models-trust-change-detail-marketing-event.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Trust Score Dashboard API
 * 기업별 신뢰 점수 시계열 데이터 API
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * Marketing Event - MBM 이벤트
 * @export
 * @interface ModelsTrustChangeDetailMarketingEvent
 */
export interface ModelsTrustChangeDetailMarketingEvent {
    /**
     * 
     * @type {string}
     * @memberof ModelsTrustChangeDetailMarketingEvent
     */
    'date': string;
    /**
     * 
     * @type {string}
     * @memberof ModelsTrustChangeDetailMarketingEvent
     */
    'title': string;
    /**
     * 
     * @type {string}
     * @memberof ModelsTrustChangeDetailMarketingEvent
     */
    'event_url'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof ModelsTrustChangeDetailMarketingEvent
     */
    'product'?: string | null;
    /**
     * 
     * @type {Array<string>}
     * @memberof ModelsTrustChangeDetailMarketingEvent
     */
    'event_target'?: Array<string>;
    /**
     * 
     * @type {string}
     * @memberof ModelsTrustChangeDetailMarketingEvent
     */
    'event_type'?: string | null;
    /**
     * 
     * @type {number}
     * @memberof ModelsTrustChangeDetailMarketingEvent
     */
    'npsScore'?: number | null;
}